Time: 0 ms (100.00%) | Memory: 23.9 MB (23.46%) - LeetSync

+++ 637-average-of-levels-in-binary-tree/average-of-levels-in-binary-tree.cpp
@@ -23,7 +23,7 @@ public:
 
         while(!q.empty()){
             int n = q.size();
-            double sum = 0;
+            long sum = 0;
 
             for(int i=0; i<n; i++){
                 TreeNode* node = q.front();
@@ -40,7 +40,7 @@ public:
                 }
             }
 
-            ans.push_back(sum / n);
+            ans.push_back((double)sum / n);
         }
         return ans;
     }
